+++ azalea-protocol/src/packets/game/c_update_attributes.rs
@@ -1,8 +1,8 @@
 use azalea_buf::AzBuf;
+use azalea_core::entity_id::MinecraftEntityId;
 use azalea_inventory::components::AttributeModifier;
 use azalea_protocol_macros::ClientboundGamePacket;
 use azalea_registry::builtin::Attribute;
-use azalea_core::entity_id::MinecraftEntityId;
 
 #[derive(AzBuf, ClientboundGamePacket, Clone, Debug, PartialEq)]
 pub struct ClientboundUpdateAttributes {